# backend/swagger/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/resources/view:
    get:
      consumes:
      - application/json
      description: Returns raw file bytes for inline UI viewing. Requires a viewToken
        minted by GET /resources. Never counts toward download limits or activity
        and does not require download permission.
      parameters:
      - description: Source name for the file (required)
        in: query
        name: source
        required: true
        type: string
      - description: File path
        in: query
        name: file
        required: true
        type: string
      - description: Opaque view grant token from file metadata
        in: query
        name: viewToken
        required: true
        type: string
      responses:
        "200":
          description: Raw file content (inline)
          schema:
            type: file
        "403":
          description: Missing or invalid view token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: File not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: View content of a single non-media file inline
      tags:
      - Resources
...
  /public/api/resources/view:
    get:
      consumes:
      - application/json
      description: Returns raw file bytes for inline UI viewing on a share link. Requires
        viewToken from GET /public/api/resources. Does not count toward download limits.
      parameters:
      - description: Share hash for authentication
        in: query
        name: hash
        required: true
        type: string
      - description: File path within the share
        in: query
        name: file
        required: true
        type: string
      - description: Opaque view grant token from share file metadata
        in: query
        name: viewToken
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Raw file content (inline)
          schema:
            type: file
        "403":
          description: Missing or invalid view token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Share or file not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: View a single non-media file from a public share inline
      tags:
      - Resources
...
swagger: "2.0"